requirements.txt
mediapipe==0.10.9
opencv-python-headless==4.8.1.78
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.1.0+cpu
torchvision==0.16.0+cpu
fastapi==0.109.0
uvicorn==0.27.0
gunicorn==21.2.0
python-multipart==0.0.6
Pillow==10.0.0
numpy==1.24.3
python-dotenv==1.0.0 